Move ``ensure`` definition from utils to exceptions
