+++ src/nacl/exceptions.py
@@ -45,3 +45,24 @@ class ValueError(ValueError, CryptoError):
 
 class InvalidkeyError(CryptoError):
     pass
+
+
+def ensure(cond, *args, **kwds):
+    """
+    Return if a condition is true, otherwise raise a caller-configurable
+    :py:class:`Exception`
+    :param bool cond: the condition to be checked
+    :param sequence args: the arguments to be passed to the exception's
+                          constructor
+    The only accepted named parameter is `raising` used to configure the
+    exception to be raised if `cond` is not `True`
+    """
+    _CHK_UNEXP = 'check_condition() got an unexpected keyword argument {0}'
+
+    raising = kwds.pop('raising', AssertionError)
+    if kwds:
+        raise TypeError(_CHK_UNEXP.format(repr(kwds.popitem()[0])))
+
+    if cond is True:
+        return
+    raise raising(*args)
