+++ src/nacl/utils.py
@@ -17,7 +17,6 @@ from __future__ import absolute_import, division, print_function
 import six
 
 import nacl.bindings
-from nacl import exceptions as exc
 
 
 class EncryptedMessage(six.binary_type):
@@ -66,24 +65,3 @@ def bytes_as_string(bytes_in):
 
 def random(size=32):
     return nacl.bindings.randombytes(size)
-
-
-def ensure(cond, *args, **kwds):
-    """
-    Return if a condition is true, otherwise raise a caller-configurable
-    :py:class:`Exception`
-    :param bool cond: the condition to be checked
-    :param sequence args: the arguments to be passed to the exception's
-                          constructor
-    The only accepted named parameter is `raising` used to configure the
-    exception to be raised if `cond` is not `True`
-    """
-    _CHK_UNEXP = 'check_condition() got an unexpected keyword argument {0}'
-
-    raising = kwds.pop('raising', AssertionError)
-    if kwds:
-        raise exc.TypeError(_CHK_UNEXP.format(repr(kwds.popitem[0])))
-
-    if cond is True:
-        return
-    raise raising(*args)
